check additional parentage when looking for a JPO
On newer Mechanipus games like DarkSpires, upstream code like Anomaly
Jobs is re-parented so that local changes are separate in the mush db,
making future updates easier.  Unfortunately, this requires a more lax
check or a more clever check.  Here, we've opted for a more clever
check.  The parent calls are necessary because jobs have parent() of
their bucket, and buckets have parent() of their JPO.  On DarkSpires,
the real JPO with original AJ install name is the parent of the acting
JPO.

There is still a nasty OaOO violation in that both fn_post_job and
trig_post_job duplicate a lot of these same checks and implementation.